Refactor application to submodules.
This is a NON WORKING commit.

This is a big set of changes which refactor the application into defined
modules, aiming for some logical split. This will help maintainability,
but should also allow us to fix the issues with circular imports (which mean
things need to be imported inside functions).

There are still some areas where formatting is weird, and uncommented methods,
etc. These will be fixed once it's working.