Do not create model children of items.
The current implementation creates models for each set of data. This isn't
necessary -- we can use a single model and replace the data (e.g. for a
view). In this commit we've removed the creation of models for each element
in the tree which saves a bit of time on load.

Instead, we store the relevant data as simple Python lists, and interact
using list indexing. This also simplifies any processing/import/export code.

This same data is set onto the models when viewing. This is done on the same
object (not a copy) so any interactions in the table view will apply back to
the simple list.
.